Add Netbeans GUI
This adds a .java file with the Netbeans GroupForm GUI builder version of the drawing interface, which now includes a semi-functional history interface that only tracks posts you have submitted yourself. However, the method "convertToBI(JPanel jp)" is not currently functional and regardless of what is drawn, the BufferedImage added to the HistoryScrollPane is always blank. Figured you guys might still like to see what progress I've made with the GUI. It's kind of held together with popsicle sticks and gum at this point, I will try to make it all more efficient and elegantly implemented later.